Support padding hostgroup queues when executing actions
Consider an action which is configured to run on two hostgroups of different sizes -- one large and one small. For simplicity, assume the large hostgroup has twice as many hosts as the small hostgroup. Without padding, the action will be exected on large[0] and small[0] simultaneously, then large[1] and small[1], then large[2] and small[2], etc. The smaller hostgroup will be finished halfway through the execution of the large hostgroup. If the smaller hostgroup is configured with padding of 1, the action will be run on large[0] and small[0] simultaneously, then large[1], then large[2] and small[1], etc. The smaller hostgroup would therefore finish at approximately the same time. If the smaller hostgroup is only one third the size of the larger hostgroup, you would use padding of 2 to have the groups finish at the same time.